+++ chopper/CHANGELOG.md
@@ -10,8 +10,10 @@
   }
   ```
 - Remove deprecated `Chopper.service<Type>(Type)`
+- Add `QueryMap` annotation
 - Fix https://github.com/lejard-h/chopper/issues/28
 - Fix https://github.com/lejard-h/chopper/issues/21
+- Fix https://github.com/lejard-h/chopper/issues/37
 
 ## 2.2.0